+++ src/db-toolkit/main.py
@@ -35,7 +35,7 @@ from ws.migrator import websocket_migrator
 app = FastAPI(
     title="DB Toolkit API",
     description="Database management toolkit API",
-    version="0.3.0",
+    version="0.4.0",
     swagger_ui_parameters={"defaultModelsExpandDepth": -1},
     lifespan=lifespan,
 )